style: 全仓 forge fmt（对齐 CI 的 forge 1.7.1）
仓库此前从未通过 CI 的 forge fmt --check（main 即红），本次以 CI 同版本
forge 1.7.1 全量格式化，消除 CI 格式检查失败。

Ref: #5, #6

+++ src/interfaces/IPyth.sol
@@ -6,7 +6,7 @@ interface IPyth {
         int64 price;
         uint64 conf;
         int32 expo;
-        uint publishTime;
+        uint256 publishTime;
     }
 
     function getPriceUnsafe(bytes32 id) external view returns (Price memory price);
